+++ bambi/backends/__init__.py
@@ -1,0 +1,5 @@
+from .pymc3 import PyMC3BackEnd
+
+__all__ = [
+    'PyMC3BackEnd'
+]
